// src/app/types.tsx
export interface movieObjType {
    title: string;
    date: string;
    language: string;
    actors: string;
    country: string;
    rating: string;
    trailer: string;
}
export interface updateMovieObjType {
    title: string;
    date: string;
    language: string;
    actors: string;
    country: string;
    rating: string;
    trailer: string;
}